# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  message: string;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  admin: "admin",
  reviewer: "reviewer",
  volunteer: "volunteer",
  vulnerable: "vulnerable",
  nok: "nok",
} as const;

/**
 * Roles that can be administered via /admin/users (NOK is auto-managed via its linked Vulnerable, hence excluded).
 */
export type ManagedUserRole =
  (typeof ManagedUserRole)[keyof typeof ManagedUserRole];

export const ManagedUserRole = {
  admin: "admin",
  reviewer: "reviewer",
  volunteer: "volunteer",
  vulnerable: "vulnerable",
} as const;

export type SignupRole = (typeof SignupRole)[keyof typeof SignupRole];

export const SignupRole = {
  volunteer: "volunteer",
  vulnerable: "vulnerable",
} as const;

export interface VolunteerProfileInput {
  skills?: string;
  gpsConsent: boolean;
}

export interface VulnerableProfileInput {
  address: string;
  nokName: string;
  nokRelation: string;
  nokContact: string;
}

/**
 * Optional NOK login credentials when registering a Vulnerable. When supplied, a linked nok_users row is created so the next-of-kin can log in and receive SOS notifications.
 */
export interface NokAccountInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface SignupRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  /** @minItems 1 */
  roles: SignupRole[];
  volunteer?: VolunteerProfileInput;
  vulnerable?: VulnerableProfileInput;
  nokAccount?: NokAccountInput;
}

export interface LoginRequest {
  email: string;
  password: string;
  role?: Role;
}

export type EmailVerifiedRole =
  (typeof EmailVerifiedRole)[keyof typeof EmailVerifiedRole];

export const EmailVerifiedRole = {
  volunteer: "volunteer",
  vulnerable: "vulnerable",
} as const;

export interface EmailVerified {
  ok: boolean;
  email: string;
  role: EmailVerifiedRole;
}

export type AdminRole = (typeof AdminRole)[keyof typeof AdminRole];

export const AdminRole = {
  reviewer: "reviewer",
  admin: "admin",
} as const;

export interface AdminCreateUserRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  role: AdminRole;
}

/**
 * All fields optional. Only provided fields are updated. Per-role extras are ignored when not applicable.
 */
export interface AdminUpdateUserRequest {
  name?: string;
  email?: string;
  /** @minLength 8 */
  password?: string;
  /** Volunteer only. */
  skills?: string | null;
  /** Vulnerable only. */
  address?: string;
  /** Vulnerable only. */
  nokName?: string;
  /** Vulnerable only. */
  nokRelation?: string;
  /** Vulnerable only. */
  nokContact?: string;
}

export interface AdminCreatedUser {
  id: number;
  email: string;
  name: string;
  role: AdminRole;
}

export interface AdminListAdmin {
  id: number;
  email: string;
  name: string;
  disabled: boolean;
  createdAt: string;
}

export interface AdminListReviewer {
  id: number;
  email: string;
  name: string;
  disabled: boolean;
  createdAt: string;
}

export interface AdminListVolunteer {
  id: number;
  email: string;
  name: string;
  disabled: boolean;
  skills?: string | null;
  gpsConsent: boolean;
  emailVerified: boolean;
  verified: boolean;
  lastSeenAt?: string | null;
  createdAt: string;
}

export interface AdminListVulnerable {
  id: number;
  email: string;
  name: string;
  disabled: boolean;
  verified: boolean;
  emailVerified: boolean;
  address: string;
  nokName: string;
  nokRelation: string;
  nokContact: string;
  createdAt: string;
}

export interface ReviewerUsersOverview {
  volunteers: AdminListVolunteer[];
  vulnerables: AdminListVulnerable[];
}

export interface AdminUsersOverview {
  admins: AdminListAdmin[];
  reviewers: AdminListReviewer[];
  volunteers: AdminListVolunteer[];
  vulnerables: AdminListVulnerable[];
}

export type EmergencyType = (typeof EmergencyType)[keyof typeof EmergencyType];

export const EmergencyType = {
  minor: "minor",
  major: "major",
} as const;

export type EmergencyStatus =
  (typeof EmergencyStatus)[keyof typeof EmergencyStatus];

export const EmergencyStatus = {
  active: "active",
  deactivated: "deactivated",
  resolved: "resolved",
} as const;

export type EmergencyMyResponse =
  | (typeof EmergencyMyResponse)[keyof typeof EmergencyMyResponse]
  | null;

export const EmergencyMyResponse = {
  accepted: "accepted",
  declined: "declined",
} as const;

export interface EmergencyResponseStats {
  accepted: number;
  declined: number;
  arrived: number;
  total: number;
}

export type EmergencyResponderStatus =
  (typeof EmergencyResponderStatus)[keyof typeof EmergencyResponderStatus];

export const EmergencyResponderStatus = {
  accepted: "accepted",
  declined: "declined",
} as const;

export interface EmergencyResponder {
  volunteerId: number;
  name: string;
  status: EmergencyResponderStatus;
  respondedAt: string;
  arrivedAt?: string | null;
  /** Current straight-line distance from volunteer's last GPS to the emergency. */
  distanceM?: number | null;
  /** Estimated time to arrival (walking 5 km/h). Null if no GPS or already arrived. */
  etaSeconds?: number | null;
}

export interface Emergency {
  id: number;
  type: EmergencyType;
  status: EmergencyStatus;
  creatorRole: Role;
  creatorUserId: number;
  creatorName: string;
  description?: string | null;
  lat?: number | null;
  lng?: number | null;
  address?: string | null;
  createdAt: string;
  deactivatedAt?: string | null;
  /** Volunteer-relative distance in metres, only present in volunteer scope. */
  distanceM?: number | null;
  myResponse?: EmergencyMyResponse;
  /** When the current Volunteer marked themselves arrived (volunteer scope only). */
  myArrivedAt?: string | null;
  responseStats?: EmergencyResponseStats | null;
  /** Per-volunteer responder rows. Reviewer/Admin scope only. */
  responders?: EmergencyResponder[] | null;
}

export interface EmergencyList {
  emergencies: Emergency[];
}

export interface CreateEmergencyRequest {
  type: EmergencyType;
  description?: string;
  lat?: number;
  lng?: number;
  address?: string;
}

export type RespondEmergencyRequestStatus =
  (typeof RespondEmergencyRequestStatus)[keyof typeof RespondEmergencyRequestStatus];

export const RespondEmergencyRequestStatus = {
  accepted: "accepted",
  declined: "declined",
} as const;

export interface RespondEmergencyRequest {
  status: RespondEmergencyRequestStatus;
}

export interface PendingVulnerable {
  id: number;
  email: string;
  name: string;
  address: string;
  nokName: string;
  nokRelation: string;
  nokContact: string;
  emailVerified: boolean;
  createdAt: string;
}

export interface PendingVulnerableList {
  items: PendingVulnerable[];
}

export type NokMeLinkedVulnerable = {
  id: number;
  name: string;
  address: string;
  verified: boolean;
  lastLat: number | null;
  lastLng: number | null;
  lastSeenAt: string | null;
};

export interface NokMe {
  id: number;
  name: string;
  email: string;
  contact: string;
  linkedVulnerable: NokMeLinkedVulnerable;
}

export interface VulnerableMe {
  id: number;
  name: string;
  email: string;
  address: string;
  nokName: string;
  nokRelation: string;
  nokContact: string;
  verified: boolean;
  emailVerified: boolean;
}

export interface UpdateLocationRequest {
  lat: number;
  lng: number;
}

export interface SessionUser {
  id: number;
  email: string;
  name: string;
  role: Role;
  verified?: boolean;
  emailVerified: boolean;
}

export type VerifyEmailParams = {
  token: string;
  role: VerifyEmailRole;
};

export type VerifyEmailRole =
  (typeof VerifyEmailRole)[keyof typeof VerifyEmailRole];

export const VerifyEmailRole = {
  volunteer: "volunteer",
  vulnerable: "vulnerable",
} as const;