# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Sign up as Volunteer and/or Vulnerable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered for the chosen role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to a specific role vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/verify:
    get:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify an email address using the token sent in the signup email
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
        - in: query
          name: role
          required: true
          schema:
            type: string
            enum: [volunteer, vulnerable]
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailVerified"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend the verification email for the signed-in user
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Already verified or wrong role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current session user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Role:
      type: string
      enum: [admin, reviewer, volunteer, vulnerable, nok]
...
    SignupRole:
      type: string
      enum: [volunteer, vulnerable]
    VolunteerProfileInput:
      type: object
      properties:
        skills:
          type: string
        gpsConsent:
          type: boolean
      required:
        - gpsConsent
    VulnerableProfileInput:
      type: object
      properties:
        address:
          type: string
        nokName:
          type: string
        nokRelation:
          type: string
        nokContact:
          type: string
      required:
        - address
        - nokName
        - nokRelation
        - nokContact
    NokAccountInput:
      type: object
      description: Optional NOK login credentials when registering a Vulnerable. When supplied, a linked nok_users row is created so the next-of-kin can log in and receive SOS notifications.
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    SignupRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
        roles:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SignupRole"
        volunteer:
          $ref: "#/components/schemas/VolunteerProfileInput"
        vulnerable:
          $ref: "#/components/schemas/VulnerableProfileInput"
        nokAccount:
          $ref: "#/components/schemas/NokAccountInput"
      required:
        - email
        - password
        - name
        - roles
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          $ref: "#/components/schemas/Role"
      required:
        - email
        - password
    EmailVerified:
      type: object
      properties:
        ok:
          type: boolean
        email:
          type: string
        role:
          type: string
          enum: [volunteer, vulnerable]
      required:
        - ok
        - email
        - role
...
    SessionUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        verified:
          type: boolean
        emailVerified:
          type: boolean
      required:
        - id
        - email
        - name
        - role
        - emailVerified