# lib/api-zod/src/generated/types/nokAccountInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Optional NOK login credentials when registering a Vulnerable. When supplied, a linked nok_users row is created so the next-of-kin can log in and receive SOS notifications.
 */
export interface NokAccountInput {
  email: string;
  /** @minLength 8 */
  password: string;
}