# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  admin: "admin",
  reviewer: "reviewer",
  volunteer: "volunteer",
  vulnerable: "vulnerable",
  nok: "nok",
} as const;