# lib/api-zod/src/generated/types/signupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NokAccountInput } from "./nokAccountInput";
import type { SignupRole } from "./signupRole";
import type { VolunteerProfileInput } from "./volunteerProfileInput";
import type { VulnerableProfileInput } from "./vulnerableProfileInput";

export interface SignupRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name: string;
  /** @minItems 1 */
  roles: SignupRole[];
  volunteer?: VolunteerProfileInput;
  vulnerable?: VulnerableProfileInput;
  nokAccount?: NokAccountInput;
}